Move remove_injected_params_from_signature and annotation_context into dependency_injection.py
